Split Reportwriter into classes specialized for export types
Separates out a lot of unrelated code.
